+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 # 1.7.1
 * renamed commands. Now they are `thumber.clear` and `thumber.clear_all`;
 * requires `me-tools` package for dev;
+* removed `ConsoleIntegrationTestTrait`, because it is now sufficient to use the
+    same trait provided by `me-tools`;
 * updated for CakePHP 3.7.1 and `php-tools` 1.1.12.
 
 # 1.7.0